Update plot1.R
I updated the first line from each file to use read.table instead of read.csv since the R script is reading from a .txt file not a .csv file. In practice, both commands function equally well in reading the .txt file into the script but since we are not to run the scripts while reviewing each others' work I made this change for clarity.